refactor(summarizer): restructure project for streamlined summarization
