--- modules/github/dagger.json
@@ -1,5 +1,0 @@
-{
-  "name": "github",
-  "sdk": "go",
-  "source": "."
-}
