--- modules/markdown/dagger.json
@@ -1,5 +1,0 @@
-{
-  "name": "markdown",
-  "sdk": "go",
-  "source": "."
-}
