--- modules/summarizer/dagger.json
@@ -1,5 +1,0 @@
-{
-  "name": "summarizer",
-  "sdk": "go",
-  "source": "."
-}
