+++ summarizer/dagger.json
@@ -1,0 +1,7 @@
+{
+  "name": "summarizer",
+  "engineVersion": "v0.18.8",
+  "sdk": {
+    "source": "go"
+  }
+}
